[WIP] breakup the overcloud playbooks into logical steps
The goal w/ this patch is to start breaking down playbooks into independent
and idempotent steps as much as possible. The breakdown of the overcloud
should follow the tripleo docs as much as possible. Note,
overcloud/register-nodes, overcloud/introspect-nodes is terminology that
matches directly with the tripleo documentation.

In this change you will find parts of the basic and advanced deployment topics.
Another benefit of breaking the deployment playbooks into their corresponding
doc steps is that we will be able to add new features to the deployment e.g.
"Deploying Manila" or "Configuring Cinder with a netapp backend" with out
distrupting the current playbooks.

Additionally this enables a user to step through the automation along side
the documentation which should allow developers and users to retry steps,
plug in development code, or rollback a step if required.  There are many
ways to enable a step by step deployment, after considering the options we have
concluded that using a directory structure with a 1-1 mapping to the docs and
the yaml description of the doc placed directly in the playbooks offers the
greatest transparency to the reader.

Change-Id: I99d23a5dacc57163cb5e057c566de5198f941a9f